add new repository TMDB

+++ app/src/main/java/com/fenixgs/filmedodia/domain/model/Movie.kt
@@ -5,5 +5,5 @@ data class Movie(
     val title: String,
     val description: String,
     val posterUrl: String?,
-    val rating: Float
+    val rating: Double
 )
